fix: have less strict size-checks for shapes
We were requiring shapes to have a size (as written in their header)
that strictly matches a value we could expect given some parameters like
the number of points.

However this is probably a bit to strict and now if the header's
record_size is bigger than something we would expect, we will now read
and then skip the rest of the bytes, instead of returning an error.

If the header's record_size is too small, we for now still return an
error.